# api/openapi.yaml
openapi: 3.0.3
info:
  title: Simplus API
  version: 0.1.0
  description: Locale-neutral control-plane contract for Simplus.
servers:
  - url: /
paths:
...
  /api/v1/inventory:
    get:
      operationId: getInventory
      summary: Read the discovered device and active-line inventory
      tags: [inventory]
      responses:
        "200":
          description: Current inventory snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "409":
          description: The instance has not completed setup or is not ready for business APIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "421":
          description: The HTTP authority is not a permitted loopback host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: The inventory snapshot could not be produced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/v1/euicc:
    get:
      operationId: getEUICCState
      summary: Read Simulator removable-eUICC installed profiles
      tags: [euicc]
      responses:
        "200": { description: Simulator eUICC state, content: { application/json: { schema: { $ref: "#/components/schemas/EUICCState" } } } }
        "401": { description: No administrator session, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "409": { description: Instance not ready, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "503": { description: eUICC management unavailable in this runtime, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
  /api/v1/euicc/profiles/{profileId}/activate:
    parameters:
      - name: profileId
        in: path
        required: true
        schema: { type: string, pattern: "^simulator-euicc-profile-[ab]$" }
    post:
      operationId: activateEUICCProfile
      summary: Switch active profile in Simulator and read back confirmation
      tags: [euicc]
      responses:
        "200": { description: Switched and confirmed state, content: { application/json: { schema: { $ref: "#/components/schemas/EUICCState" } } } }
        "400": { description: Invalid profile, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "401": { description: No administrator session, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "404": { description: Profile not found, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "409": { description: Instance not ready, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "503": { description: eUICC management unavailable in this runtime, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
...
  /api/v1/hardware/topology:
    get:
      operationId: getHardwareTopology
      summary: Read the normalized dynamic hardware, SIM, Profile, ResourceGroup, and Line graph
      tags: [inventory]
      responses:
        "200":
          description: Current normalized hardware topology
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HardwareTopologyResponse"
        "401":
          description: No live administrator session is present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The instance has not completed setup or is not ready for business APIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "421":
          description: The HTTP authority is not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: The hardware topology could not be produced or failed validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "504":
          description: The request exceeded the bounded control-plane deadline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/v1/modems:
    get:
      operationId: listManagedModems
      summary: List administrator-added modems and their current observed state
      tags: [inventory]
      responses:
        "200":
          description: Managed modem list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedModemList" }
        "401":
          description: No live administrator session is present
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "409":
          description: The instance is not ready for business APIs
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "500":
          description: Managed modems could not be read
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    post:
      operationId: addManagedModem
      summary: Add one currently discovered modem without changing hardware state
      tags: [inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddManagedModemRequest" }
      responses:
        "201":
          description: Modem added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedModem" }
        "400":
          description: Candidate selection is invalid
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "401":
          description: No live administrator session is present
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "404":
          description: Candidate is no longer present
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "409":
          description: Candidate is already managed, not ready, or has a conflicting equipment identity
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "500":
          description: Managed modem could not be persisted
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /api/v1/modem-candidates:
    get:
      operationId: listModemCandidates
      summary: Scan and list currently discovered modems that have not been added
      tags: [inventory]
      responses:
        "200":
          description: Current unadded modem candidates
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModemCandidateList" }
        "401":
          description: No live administrator session is present
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "409":
          description: The instance is not ready for business APIs
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "500":
          description: Current hardware candidates could not be scanned
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /api/v1/lines:
    get:
      operationId: listManagedLines
      summary: List administrator-created Lines and their resolved runtime state
      tags: [lines]
      responses:
        "200": { description: Managed Line list, content: { application/json: { schema: { $ref: "#/components/schemas/ManagedLineList" } } } }
        "401": { description: No live administrator session is present, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "409": { description: The instance is not ready for business APIs, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "500": { description: Managed Lines could not be read, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
    post:
      operationId: addManagedLine
      summary: Create only a stable Line identity from one current managed-modem SIM/Profile candidate
      description: Revalidates the transient candidate and persists its immutable binding plus display name; does not change RF, Mihomo, Host VoWiFi, SMS, or calls.
      tags: [lines]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/AddManagedLineRequest" } } }
      responses:
        "201": { description: Line created, content: { application/json: { schema: { $ref: "#/components/schemas/ManagedLine" } } } }
        "400": { description: "Candidate or name is invalid", content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "401": { description: No live administrator session is present, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "404": { description: Candidate is no longer present, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "409": { description: Candidate is already represented by a Line, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "500": { description: Managed Line could not be persisted, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
  /api/v1/line-candidates:
    get:
      operationId: listLineCandidates
      summary: List current Line candidates and typed readiness for every administrator-added modem
      description: Candidate IDs are transient observations. Responses contain bounded display metadata and capabilities, never SIM identity fingerprints, runtime hardware targets, sysfs paths, or device nodes.
      tags: [lines]
      responses:
        "200": { description: Current addable and unavailable Line candidates, content: { application/json: { schema: { $ref: "#/components/schemas/LineCandidateList" } } } }
        "401": { description: No live administrator session is present, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "409": { description: The instance is not ready for business APIs, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "500": { description: Current Line candidates could not be resolved, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
  /api/v1/lines/{lineId}:
    parameters:
      - { name: lineId, in: path, required: true, schema: { type: string, pattern: "^line_[A-Za-z0-9_-]{22}$" } }
    put:
      operationId: updateManagedLine
      summary: Update a Line display name without rebinding hardware
      tags: [lines]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateManagedLineRequest" } } }
      responses:
        "200": { description: Line updated, content: { application/json: { schema: { $ref: "#/components/schemas/ManagedLine" } } } }
        "400": { description: Name is invalid, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "401": { description: No live administrator session is present, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "404": { description: Managed Line was not found, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "500": { description: Managed Line could not be updated, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
  /api/v1/modems/{modemId}/rf-state:
    put:
      operationId: setManagedModemRFState
      summary: Set one managed modem's runtime RF state through its typed adapter
      tags: [inventory]
      parameters:
        - name: modemId
          in: path
          required: true
          schema: { type: string, pattern: "^modem_[A-Za-z0-9_-]{22}$" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetManagedModemRFStateRequest" }
      responses:
        "200":
          description: RF state read back and confirmed
          content: { application/json: { schema: { $ref: "#/components/schemas/ManagedModem" } } }
        "400":
          description: RF request is invalid
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "401":
          description: No live administrator session is present
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "404":
          description: Managed modem was not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "409":
          description: RF state cannot currently be changed or confirmed
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "422":
          description: RF control is unsupported for this modem
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "500":
          description: RF operation failed
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /api/v1/modems/{modemId}/equipment-identity:
    post:
      operationId: readManagedModemEquipmentIdentity
      summary: Read one online managed modem's IMEI on explicit administrator request
      tags: [inventory]
      parameters:
        - name: modemId
          in: path
          required: true
          schema: { type: string, pattern: "^modem_[A-Za-z0-9_-]{22}$" }
      responses:
        "200":
          description: Current IMEI read directly from the resolved modem; never persisted
          headers:
            Cache-Control:
              schema: { type: string, enum: [no-store] }
          content: { application/json: { schema: { $ref: "#/components/schemas/ManagedModemEquipmentIdentity" } } }
        "401":
          description: No live administrator session is present
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "404":
          description: Managed modem was not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "409":
          description: Modem is offline or its identity changed during the read
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "422":
          description: Equipment identity read is unsupported for this modem
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "503":
          description: Equipment identity is temporarily unavailable
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
...
components:
...
  schemas:
...
    ManagedModemList:
      type: object
      additionalProperties: false
      required: [modems]
      properties:
        modems:
          type: array
          maxItems: 64
          items: { $ref: "#/components/schemas/ManagedModem" }
...
    ManagedLineList:
      type: object
      additionalProperties: false
      required: [lines]
      properties:
        lines: { type: array, maxItems: 1024, items: { $ref: "#/components/schemas/ManagedLine" } }
...
    LineCandidateList:
      type: object
      additionalProperties: false
      required: [candidates]
      properties:
        candidates: { type: array, maxItems: 1024, items: { $ref: "#/components/schemas/LineCandidate" } }
    AddManagedLineRequest:
      type: object
      additionalProperties: false
      required: [candidateId, displayName]
      properties:
        candidateId: { type: string, pattern: "^line-candidate-[0-9a-f]{32}$" }
        displayName: { type: string, minLength: 1, maxLength: 120 }
...
    ModemCandidateList:
      type: object
      additionalProperties: false
      required: [candidates]
      properties:
        candidates:
          type: array
          maxItems: 64
          items: { $ref: "#/components/schemas/ModemCandidate" }
    AddManagedModemRequest:
      type: object
      additionalProperties: false
      required: [candidateId]
      properties:
        candidateId: { type: string, pattern: "^[a-z0-9][a-z0-9-]{0,63}$" }
...
    HardwareTopologyResponse:
      type: object
      additionalProperties: false
      required: [generation, revision, observedAt, devices, modemFunctions, simSlots, simMedia, subscriptionProfiles, resourceGroups, lines]
      properties:
        generation: { $ref: "#/components/schemas/HardwareGeneration" }
        revision: { $ref: "#/components/schemas/TopologyRevision" }
        observedAt: { type: string, format: date-time }
        devices:
          type: array
          items: { $ref: "#/components/schemas/PhysicalDeviceDetail" }
        modemFunctions:
          type: array
          items: { $ref: "#/components/schemas/ModemFunctionDetail" }
        simSlots:
          type: array
          items: { $ref: "#/components/schemas/SIMSlotDetail" }
        simMedia:
          type: array
          items: { $ref: "#/components/schemas/SIMMediaDetail" }
        subscriptionProfiles:
          type: array
          items: { $ref: "#/components/schemas/SubscriptionProfileDetail" }
        resourceGroups:
          type: array
          items: { $ref: "#/components/schemas/ResourceGroupDetail" }
        lines:
          type: array
          items: { $ref: "#/components/schemas/HardwareLineDetail" }
    InventoryResponse:
      type: object
      additionalProperties: false
      required: [generation, revision, observedAt, devices, lines]
      properties:
        generation:
          $ref: "#/components/schemas/HardwareGeneration"
        revision:
          $ref: "#/components/schemas/TopologyRevision"
        observedAt:
          type: string
          format: date-time
        devices:
          type: array
          items:
            $ref: "#/components/schemas/PhysicalDeviceSummary"
        lines:
          type: array
          items:
            $ref: "#/components/schemas/LineSummary"
...
    EUICCState:
      type: object
      additionalProperties: false
      required: [eidHint, profiles]
      properties:
        eidHint: { type: string, minLength: 1, maxLength: 80 }
        profiles:
          type: array
          minItems: 1
          maxItems: 16
          items: { $ref: "#/components/schemas/EUICCProfile" }
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, retryable]
      properties:
        code:
          type: string
          pattern: "^[A-Z0-9_]+$"
        retryable:
          type: boolean
        reference:
          type: string